refactor: get 요청도 로그 찍도록 수정
